+++ CHANGELOG.md
@@ -40,6 +40,8 @@
 - [deprecated] - `seed::browser::service::fetch` module is deprecated in favor of `seed::browser::fetch`.
 - Added example `fetch`.
 - Implemented `IntoNodes` for `Option<Node<Msg>>` and `Option<Vec<Node<Msg>>>`.
+- Added example `graphql` (#400).
+- Implemented `UpdateEl` for `i64` and `u64`.
 
 ## v0.6.0
 - Implemented `UpdateEl` for `Filter` and `FilterMap`.